requirements.txt
langchain
langchain-community
langchain-openai
langgraph
pydantic
faiss-cpu
python-dotenv
tiktoken
vllm
